# Conflicts:
#	src/main/java/com/example/eksamensprojektbilabonnement/services/CustomerService.java

+++ src/main/java/com/example/eksamensprojektbilabonnement/services/CustomerService.java
@@ -1,7 +1,9 @@
 package com.example.eksamensprojektbilabonnement.services;
 
 import com.example.eksamensprojektbilabonnement.models.Customer;
+import com.example.eksamensprojektbilabonnement.models.LeaseAgreement;
 import com.example.eksamensprojektbilabonnement.repositories.CustomerRepository;
+import com.example.eksamensprojektbilabonnement.repositories.LeaseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
@@ -14,23 +16,36 @@ public class CustomerService {
     @Autowired
     CustomerRepository customerRepository;
 
+    @Autowired
+    LeaseRepository leaseRepository;
+
     public List<Customer> getAllCustomers() {
         return customerRepository.getAllCustomers();
     }
 
-    public String deleteCustomer(int customerId) {
-        try {
-            customerRepository.deleteCustomer(customerId);
-        } catch (Exception e) {
-            if (e.getCause() instanceof SQLIntegrityConstraintViolationException) {
-                customerRepository.anonymizeCustomerData(customerId);
-                return "This customer has active leases, and cannot be deleted. Customer data has been anonymized";
-            }
-        } return "Customer and all its data have been deleted.";
-    }
-
     public List<Customer> getNonAnonymousCustomers() {
         return customerRepository.getNonAnonymousCustomers();
+    }
+
+    public String deleteCustomer(int customerId) {
+        //Checks if there are any non concluded leases for the customer:
+        List<LeaseAgreement> nonConcludedLeases = leaseRepository.getNonConcludedLeases(customerId);
+
+        //Tries to delete the customer from the database, if there are no non concluded leases.
+        // If there are foreign key constraints, the data is anonymized instead
+        if (nonConcludedLeases.isEmpty()) {
+            try {
+                customerRepository.deleteCustomer(customerId);
+            } catch (Exception e) {
+                if (e.getCause() instanceof SQLIntegrityConstraintViolationException) {
+                    customerRepository.anonymizeCustomerData(customerId);
+                    return "This customer has stored leases, and cannot be deleted. Customer data has been anonymized";
+                }
+            }
+            return "Customer and all its data have been deleted.";
+        } else {
+            return "Customer has non concluded leases, and cannot be deleted";
+        }
     }
 
     public List<Integer> findCustomersForAnonymization(){
